Change name of appendStringInfoVA to appendStringInfoVA_c.
There is the same name function in PostgreSQL server side,
but the internal side function has an additional feature;
performing buffer extension. But, this is very confusing.
This fix change the internal function name in order to
find this is client side function.

The patch was created by Amit Langote.